# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puraani Jeans APIs",
        "description": "Puraani Jeans API documentation",
        "contact": {
            "email": "developer@puranijeans.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": ""
    },
    "servers": [
        {
            "url": "http://safeblockcom.test/",
            "description": "local"
        }
    ],
    "paths": {
...
        "/v1/closet/image-upload": {
            "post": {
                "tags": [
                    "Closet"
                ],
                "summary": "Closet Icon and Banner Avatar",
                "operationId": "imageUpload",
                "requestBody": {
                    "description": "Icon and Banner Upload for a Closet",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "banner": {
                                        "description": "base64 ",
                                        "type": "string",
                                        "example": "base64:"
                                    },
                                    "icon": {
                                        "description": "base64 ",
                                        "type": "string",
                                        "example": "base64:"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/v1/add/product": {
            "post": {
                "tags": [
                    "Closet"
                ],
                "summary": "Add New Products",
                "operationId": "addProduct",
                "requestBody": {
                    "description": "Add New Products",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "filters": {
                                        "description": "Network Id of customer's phone number",
                                        "type": "object",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
        "/v1/filter/closet/{slug}/product": {
            "post": {
                "tags": [
                    "Closet"
                ],
                "summary": "Store Products",
                "operationId": "getFilteredClosetProducts",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "store slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Resend OTP",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "filters": {
                                        "description": "Network Id of customer's phone number",
                                        "type": "object",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
        "/v1/closet/{slug}/product": {
            "get": {
                "tags": [
                    "Closet"
                ],
                "summary": "Store Products",
                "operationId": "getClosetProducts",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "store slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
        "/v1/closet/{slug}/category/{catSlug}/product": {
            "get": {
                "tags": [
                    "Closet"
                ],
                "summary": "Manage Closet Category Products",
                "operationId": "getClosetCategoryProducts",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "catSlug",
                        "in": "path",
                        "description": "category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "user_access_token": {
                "type": "apiKey",
                "description": "user access token",
                "name": "Authorization",
                "in": "header"
            },
            "client_id": {
                "type": "apiKey",
                "description": "client id",
                "name": "x-client-id",
                "in": "header"
            },
            "client_secret": {
                "type": "apiKey",
                "description": "client secret",
                "name": "x-client-token",
                "in": "header"
            }
        }
    },
...
}